create: new component for messages area input
