+++ src/app/components/devspace-message/devspace-message.component.ts
@@ -1,23 +1,20 @@
-import { query } from '@angular/animations';
+
 import { Component, ElementRef, ViewChild, ChangeDetectorRef, } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import { MessageInputAreaComponent } from '@components/message-input-area/message-input-area.component';
 import { DevspaceService } from '@shared/services/devspace-service/devspace.service';
 
 @Component({
   selector: 'app-devspace-message',
-  imports: [FormsModule],
+  imports: [FormsModule,MessageInputAreaComponent],
   templateUrl: './devspace-message.component.html',
   styleUrl: './devspace-message.component.scss'
 })
 export class DevspaceMessageComponent {
-  @ViewChild('messageInput') messageInput!: ElementRef;
+  
   @ViewChild('messageTitle') messageTitle!: ElementRef;
   @ViewChild('addInput') addInput!: ElementRef;
-  openSmileyBar = false;
-  openContactBar = false;
-  openChannelBar = false;
-  openContactBarSearch = false;
-  openChannelBarSearch = false;
+  
   messageText: string = '';
 
   constructor(public devspaceService: DevspaceService, private cdRef: ChangeDetectorRef) {
@@ -32,110 +29,13 @@ export class DevspaceMessageComponent {
     }
   }
 
-  selectSmiley(i: number) {
-    const emoji = this.devspaceService.emojis[i];
-    const messageDiv = this.messageInput.nativeElement as HTMLDivElement;
-    messageDiv.focus();
-    const selection = window.getSelection();
-    const range = document.createRange();
-    if (!messageDiv.textContent?.trim()) {
-      messageDiv.innerHTML = "";
-    }
-
-    range.selectNodeContents(messageDiv);
-    range.collapse(false);
-    selection?.removeAllRanges();
-    selection?.addRange(range);
-    const textNode = document.createTextNode(emoji);
-    range.insertNode(textNode);
-    range.setStartAfter(textNode);
-    range.collapse(true);
-    selection?.removeAllRanges();
-    selection?.addRange(range);
-    this.openSmileyBar = false;
-    this.cdRef.detectChanges();
-  }
-  openAt() {
-    this.openSmileyBar = false;
-    this.openContactBar = true;
-  }
-
-  selectContact(i: number) {
-    this.openContactBar = false;
-    const contactName = this.devspaceService.accounts[i].name;
-    const messageDiv = this.messageInput.nativeElement as HTMLDivElement;
-    messageDiv.focus();
-    const selection = window.getSelection();
-    if (!selection || selection.rangeCount === 0) return;
-    const range = selection.getRangeAt(0);
-    if (messageDiv.textContent?.trim() === "") {
-      range.setStart(messageDiv, 0);
-      range.setEnd(messageDiv, 0);
-    } else {
-      const lastChild = messageDiv.lastChild;
-      if (lastChild && lastChild.nodeType === Node.TEXT_NODE) {
-        range.setStart(lastChild, lastChild.textContent?.length || 0);
-        range.setEnd(lastChild, lastChild.textContent?.length || 0);
-      } else {
-        range.setStart(messageDiv, messageDiv.childNodes.length);
-        range.setEnd(messageDiv, messageDiv.childNodes.length);
-      }
-    }
-    range.deleteContents();
-    const textBeforeCursor = range.startContainer.textContent?.slice(0, range.startOffset) || "";
-    if (textBeforeCursor.endsWith("@")) {
-      const updatedText = textBeforeCursor.slice(0, -1);
-      range.startContainer.textContent = updatedText + range.startContainer.textContent?.slice(range.startOffset);
-      range.setStart(range.startContainer, updatedText.length);
-      range.setEnd(range.startContainer, updatedText.length);
-    }
-    const nameToInsert = '@' + contactName;
-    const span = document.createElement('span');
-    span.classList.add('contact');
-    span.style.color = 'var(--mail-blue)';
-    span.textContent = nameToInsert;
-    span.contentEditable = "false";
-    span.dataset['contact'] = "true";
-    range.insertNode(span);
-    const newRange = document.createRange();
-    newRange.setStartAfter(span);
-    newRange.setEndAfter(span);
-    selection.removeAllRanges();
-    selection.addRange(newRange);
-    this.cdRef.detectChanges();
-    messageDiv.focus();
-  }
-
   ngAfterViewInit() {
-    this.checkPlaceholder();
-    this.inputSearchMessages();
-    this.inputSearchAddMessage();
+    this.checkPlaceholder();    
+    this.inputSearchAddMessage();   
+    
   }
 
-  inputSearchMessages() {
-    this.messageInput.nativeElement.addEventListener('input', () => {
-      const selection = window.getSelection();
-      if (!selection || selection.rangeCount === 0) return;
-      const range = selection.getRangeAt(0);
-      const textBeforeCursor = range.startContainer.textContent?.slice(0, range.startOffset) || "";
-      if (textBeforeCursor.endsWith("@")) {
-        this.openContactBar = true;
-        this.openContactBarSearch = false;
-        this.openChannelBarSearch = false;
-      }else{
-        this.openContactBar = false;
-      }
-
-      if (textBeforeCursor.endsWith("#")) {
-        this.openChannelBar = true;
-        this.openContactBarSearch = false;
-        this.openChannelBarSearch = false;
-      } else {
-        this.openChannelBar = false;
-      }
-    });
-  }
-
+ 
   inputSearchAddMessage() {
     this.addInput.nativeElement.addEventListener('input', () => {
       const selection = window.getSelection();
@@ -143,110 +43,24 @@ export class DevspaceMessageComponent {
       const range = selection.getRangeAt(0);
       const textBeforeCursor = range.startContainer.textContent?.slice(0, range.startOffset) || "";
       if (textBeforeCursor.endsWith("@")) {
-        this.openContactBarSearch = true;
-        this.openSmileyBar = false;
-        this.openContactBar = false;
-        this.openChannelBar = false;      
+        this.devspaceService.openContactBarSearch = true;
+        this.devspaceService.openSmileyBar = false;
+        this.devspaceService.openContactBar = false;
+        this.devspaceService.openChannelBar = false;      
       } else {
-        this.openContactBarSearch = false;
+        this.devspaceService.openContactBarSearch = false;
       }
 
       if (textBeforeCursor.endsWith("#")) {
-        this.openChannelBarSearch = true;
-        this.openSmileyBar = false;
-        this.openContactBar = false;
-        this.openChannelBar = false;
+        this.devspaceService.openChannelBarSearch = true;
+        this.devspaceService.openSmileyBar = false;
+        this.devspaceService.openContactBar = false;
+        this.devspaceService.openChannelBar = false;
       } else {
-        this.openChannelBarSearch = false;
+        this.devspaceService.openChannelBarSearch = false;
       }
     });
 
-  }
-
-  openSmiley() {
-    this.openSmileyBar = !this.openSmileyBar;
-    this.openContactBar = false;
-    this.openChannelBar = false;
-  }
-
-  closeBars() {
-    this.openSmileyBar = false;
-    this.openContactBar = false;
-    this.openChannelBar = false;
-  }
-
-  selectChannel(i: number) {
-    this.openChannelBar = false;
-    const channelsName = this.devspaceService.channels[i].name;
-    const messageDiv = this.messageInput.nativeElement as HTMLDivElement;
-    messageDiv.focus();
-    const selection = window.getSelection();
-    if (!selection || selection.rangeCount === 0) return;
-    const range = selection.getRangeAt(0);
-    if (messageDiv.textContent?.trim() === "") {
-      range.setStart(messageDiv, 0);
-      range.setEnd(messageDiv, 0);
-    } else {
-      const lastChild = messageDiv.lastChild;
-      if (lastChild && lastChild.nodeType === Node.TEXT_NODE) {
-        range.setStart(lastChild, lastChild.textContent?.length || 0);
-        range.setEnd(lastChild, lastChild.textContent?.length || 0);
-      } else {
-        range.setStart(messageDiv, messageDiv.childNodes.length);
-        range.setEnd(messageDiv, messageDiv.childNodes.length);
-      }
-    }
-    range.deleteContents();
-    const textBeforeCursor = range.startContainer.textContent?.slice(0, range.startOffset) || "";
-    if (textBeforeCursor.endsWith("#")) {
-      const updatedText = textBeforeCursor.slice(0, -1);
-      range.startContainer.textContent = updatedText + range.startContainer.textContent?.slice(range.startOffset);
-      range.setStart(range.startContainer, updatedText.length);
-      range.setEnd(range.startContainer, updatedText.length);
-    }
-    const nameToInsert = '#' + channelsName;
-    const span = document.createElement('span');
-    span.classList.add('channel');
-    span.style.color = 'var(--secondary-color)';
-    span.textContent = nameToInsert;
-    span.contentEditable = "false";
-    span.dataset['channel'] = "true";
-    range.insertNode(span);
-    const newRange = document.createRange();
-    newRange.setStartAfter(span);
-    newRange.setEndAfter(span);
-    selection.removeAllRanges();
-    selection.addRange(newRange);
-    this.cdRef.detectChanges();
-  }
-
-  addMessage() {
-    const message = this.messageInput.nativeElement.textContent;
-    const channels = this.channelNameMessage();
-    const contacts = this.contactNameMessage();
-    const channelInput = this.channelNameInput();
-    const contactInput = this.contactNameInput();
-    console.log(channels, contacts, channelInput, contactInput, message);
-
-
-    this.messageInput.nativeElement.textContent = "";
-    this.addInput.nativeElement.textContent = "";
-    this.openChannelBarSearch = false;
-    this.openContactBarSearch = false;
-    this.openSmileyBar = false;
-    this.openContactBar = false;
-    this.openChannelBar = false;
-  }
-
-  contactNameMessage() {
-    const contacts = this.messageInput.nativeElement.querySelectorAll('span[data-contact="true"]');
-    const contactArray: any[] = [];
-    contacts?.forEach((span: HTMLElement) => {
-      const contactName = span.textContent?.slice(1);
-      contactArray.push(contactName);
-    });
-    const sortContacts = [...new Set(contactArray)];
-    return sortContacts;
   }
 
   contactNameInput() {
@@ -260,16 +74,7 @@ export class DevspaceMessageComponent {
     return sortContacts;
   }
 
-  channelNameMessage() {
-    const channels = this.messageInput.nativeElement.querySelectorAll('span[data-channel="true"]');
-    const channelArray: any[] = [];
-    channels?.forEach((span: HTMLElement) => {
-      const channelName = span.textContent?.slice(1);
-      channelArray.push(channelName);
-    });
-    const sortChannels = [...new Set(channelArray)];
-    return sortChannels;
-  }
+ 
 
   channelNameInput() {
     const channels = this.addInput.nativeElement.querySelectorAll('span[data-channel-Input="true"]');
@@ -283,7 +88,7 @@ export class DevspaceMessageComponent {
   }
 
   selectContactSearch(i: number) {
-    this.openContactBarSearch = false;
+    this.devspaceService.openContactBarSearch = false;
     const contactName = this.devspaceService.accounts[i].name;
     const messageDiv = this.addInput.nativeElement as HTMLDivElement;
     const selection = window.getSelection();
@@ -333,11 +138,8 @@ export class DevspaceMessageComponent {
     this.cdRef.detectChanges();
   }
 
-
-
-
   selectChannelSearch(i: number) {
-    this.openChannelBarSearch = false;
+    this.devspaceService.openChannelBarSearch = false;
     const channelName = this.devspaceService.channels[i].name;
     const messageDiv = this.addInput.nativeElement as HTMLDivElement;
     const selection = window.getSelection();
