+++ src/app/components/message-input-area/message-input-area.component.scss
@@ -1,0 +1,230 @@
+.devspace-message-write {
+    padding-left: 71px;
+    padding-right: 71px;
+    position: relative;
+}
+.devspace-message-write-area {
+    max-width: 1352px;
+    width: 100%;
+    height: 145px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    gap: 10px;
+    padding: 20px 20px 20px 20px;
+    flex-shrink: 0;
+    border-radius: 20px;
+    border: 1px solid var(--light-purple);
+    position: relative;
+
+}
+
+.devspace-message-write-area-input::-webkit-scrollbar-button {
+    display: none;
+}
+
+.devspace-message-write-area-input::-webkit-scrollbar {
+    width: 6px;
+}
+
+.devspace-message-write-area-input::-webkit-scrollbar-track {
+    background: #f1f1f1;
+    border-radius: 35px;
+}
+
+.devspace-message-write-area-input::-webkit-scrollbar-thumb {
+    background: var(--purple-3);
+    border-radius: 35px;
+}
+
+.devspace-message-write-area-input::-webkit-scrollbar-thumb:hover {
+    background: var(--light-purple);
+    cursor: pointer;
+}
+
+.devspace-message-write-area-input {
+    width: 100%;
+    height: 80%;
+
+    &:focus {
+        outline: none;
+    }
+
+    word-break: break-word;
+    overflow: auto;
+}
+
+
+
+.devspace-message-write-area-icons {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    height: 40px;
+
+}
+
+.devspace-message-write-area-icons-bars {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+
+    & span {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        cursor: pointer;
+        width: 40px;
+        height: 40px;
+    }
+
+    & span:nth-of-type(1) {
+
+        &:hover {
+            border-radius: 50%;
+            background-color: var(--bg-color);
+
+            & img {
+                content: url('/assets/img/smiley_hover.png');
+            }
+        }
+    }
+
+    & span:nth-of-type(2) {
+        &:hover {
+            border-radius: 50%;
+            background-color: var(--bg-color);
+
+            & img {
+                content: url('/assets/img/at_hover.png');
+            }
+        }
+    }
+
+
+}
+
+.devspace-message-write-area-icons-bars-send {
+    cursor: pointer;
+
+    &:hover {
+        & img {
+            content: url('/assets/img/send_hover.png');
+        }
+    }
+}
+
+
+
+.devspace-message-contact-bar {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    top: -360px;
+    left: 20px;
+    width: 350px;
+    height: 370px;
+    overflow: auto;
+    background-color: var(--white-color);
+    border: 1px solid var(--light-purple);
+    border-top-left-radius: 20px;
+    border-top-right-radius: 20px;
+    border-bottom-right-radius: 20px;
+    padding: 20px;
+    gap: 15px;
+    z-index: 100;
+}
+
+.devspace-message-channel-bar {
+    display: flex;
+    position: absolute;
+    flex-direction: column;
+    top: -140px;
+    left: 20px;
+    max-width: 350px;
+    min-width: 250px;
+    min-height: 150px;
+    height: fit-content;
+    overflow: auto;
+    background-color: var(--white-color);
+    border: 1px solid var(--light-purple);
+    border-top-left-radius: 20px;
+    border-top-right-radius: 20px;
+    border-bottom-right-radius: 20px;
+    padding: 20px 20px 20px 20px;
+    gap: 15px;
+    z-index: 100;
+
+    & span {
+        display: flex;
+        align-items: center;
+        width: 100%;
+        height: 33px;
+        gap: 10px;
+        font-size: var(--font-size-m);
+        font-weight: var(--font-weight-normal);
+        line-height: 100%;
+        color: var(--black-color);
+        padding-right: 20px;
+        padding-left: 20px;
+        cursor: pointer;
+
+        &:hover {
+            background-color: var(--bg-color);
+            border-radius: 35px;
+
+            & img {
+                content: url('/assets/img/tagHover.png');
+            }
+        }
+    }
+}
+
+.selected-Contact-Item {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    font-size: var(--font-size-m);
+    font-weight: var(--font-weight-normal);
+    line-height: 100%;
+    color: black;
+    padding-left: 20px;
+    cursor: pointer;
+
+    & span:nth-of-type(1) {
+        position: relative;
+
+        & span {
+            position: absolute;
+        }
+    }
+
+    &:hover {
+
+        background-color: var(--bg-color);
+        border-radius: 35px;
+    }
+}
+.devspace-message-smiley-bar {
+    position: absolute;
+    min-height: 50px;
+    width: 300px;
+    top: -280px;
+    left: 20px;
+    display: flex;
+    align-items: center;
+    padding: 10px;
+    gap: 10px;
+    z-index: 100;
+    border-top-left-radius: 20px;
+    border-top-right-radius: 20px;
+    border-bottom-right-radius: 20px;
+    background-color: var(--white-color);
+    border: 1px solid var(--light-purple);
+    flex-wrap: wrap;
+
+    & span {
+        cursor: pointer;
+    }
+}
