+++ src/styles.scss
@@ -9,6 +9,7 @@
 
 
 
+
 html {
     @include mat.theme((color: (theme-type: light,
                 primary: mat.$azure-palette,
